collapsible docks instead of scroll areas
